station.h: move Modes struct definition there. remove deps from readsb.h
