+++ convert.c
@@ -23,6 +23,7 @@
 
 #include <math.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include "readsb.h"
 
 struct converter_state {
