+++ demod_2400.c
@@ -24,6 +24,7 @@
 #include "readsb.h"
 #include "mode_ac.h"
 #include "demod_2400.h"
+#include "station.h"
 #include "fifo.h"
 #include <assert.h>
 #include <math.h>
