+++ mode_s.c
@@ -57,7 +57,7 @@
 #include "mode_ac.h"
 #include "icao_filter.h"
 #include "ais_charset.h"
-
+#include "station.h"
 #include <math.h>
 #include <string.h>
