--- receiver.h
@@ -1,6 +1,0 @@
-#ifndef READSB_RECEIVER_H
-#define READSB_RECEIVER_H
-
-void receiverPositionChanged (float lat, float lon, float alt);
-
-#endif //READSB_RECEIVER_H
