+++ sdr.c
@@ -23,6 +23,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "readsb.h"
+#include "station.h"
 #include <stdio.h>
 #include <string.h>
 #include <pthread.h>
