+++ track.c
@@ -54,9 +54,11 @@
 #include "readsb.h"
 #include "mode_ac.h"
 #include "geomag.h"
+#include "station.h"
 #include <inttypes.h>
 #include <math.h>
 #include <string.h>
+#include <stdlib.h>
 
 /* #define DEBUG_CPR_CHECKS */
